requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
slack-bolt==1.18.1
requests==2.31.0
python-dotenv==1.0.0